support buckets in AWS SIGV4-only regions
To synchronize with S3 buckets in regions that only support Signature
Version 4, SIGV4 has to be enabled in the client library via a system property
(see http://docs.aws.amazon.com/AmazonS3/latest/dev/UsingAWSSDK.html) and
requests have to be directed against an endpoint in that region.

This patch allows specifying an endpoint in the Leiningen project file. If an
endpoint is set, SIGV4 is automatically enabled.